requirements.txt
3to2==1.1.1
Babel==2.3.4
Flask==0.12
Flask-Bcrypt==0.7.1
Flask-Login==0.4.0
Flask-SQLAlchemy==2.2
Jinja2==2.9.5
MarkupSafe==1.0
RPi.GPIO==0.6.3
SQLAlchemy==1.1.6
WTForms==2.1
Werkzeug==0.12
appdirs==1.4.3
argparse==1.2.1
bcrypt==3.1.3
cffi==1.9.1
chardet==2.3.0
click==6.7
colorama==0.3.2
html5lib==0.999
itsdangerous==0.24
jsonify==0.5
meld3==1.0.2
ndg-httpsclient==0.3.2
packaging==16.8
ply==3.4
pyOpenSSL==0.13.1
pyasn1==0.1.7
pycparser==2.10
pyparsing==2.2.0
python-apt==0.9.3.12
requests==2.4.3
six==1.10.0
supervisor==3.3.1
urllib3==1.9.1
wheel==0.24.0
wsgiref==0.1.2
